+++ ecs-20140526/ChangeLog.md
@@ -1,3 +1,11 @@
+2026-06-09 Version: 7.8.2
+- Update API CreateCapacityReservation: add request parameters InstanceChargeType.
+- Update API DescribeCapacityReservations: add response parameters Body.CapacityReservationSet.$.DeliveryTime.
+- Update API DescribeCapacityReservations: add response parameters Body.CapacityReservationSet.$.UnlockedTime.
+- Update API DescribeInstances: add response parameters Body.Instances.$.MetadataOptions.InstanceMetadataTags.
+- Update API DescribeRenewalPrice: add response parameters Body.PriceInfo.RelatedPrice.
+
+
 2026-06-01 Version: 7.8.1
 - Update API CreateDeploymentSet: add request parameters Affinity.
 - Update API DescribeDeploymentSets: add response parameters Body.DeploymentSets.$.Affinity.